Add DoubleMetricsObject and rename IntMetricsObject
Introduces DoubleMetricsObject for handling double metric values and moves IntMetricsObject to its own file, renaming it from MetricsObject+Syncable.swift. This refactor separates double and integer metric logic for better clarity and maintainability.
